LayoutTests/imported/w3c/web-platform-tests/resize-observer/callback-cross-realm-report-exception-expected.txt

CONSOLE MESSAGE: Error: PASS


PASS ResizeObserver reports the exception from its callback in the callback's global object
